refactor: move Validation to protocol folder
