+++ src/main/factories/login/login-validation.spec.ts
@@ -1,6 +1,6 @@
 import { EmailValidation } from '@presentation/helpers/validators/email-validation'
 import { RequiredFieldValidation } from '@presentation/helpers/validators/required-field-validation'
-import { Validation } from '@presentation/helpers/validators/validation'
+import { Validation } from '@presentation/protocols/validation'
 import { ValidationComposite } from '@presentation/helpers/validators/validation-composite'
 import { EmailValidator } from '@presentation/protocols/email-validator.interface'
 import { makeLoginValidation } from './login-validation'
