+++ src/main/factories/login/login-validation.ts
@@ -1,6 +1,6 @@
 import { EmailValidation } from '@presentation/helpers/validators/email-validation'
 import { RequiredFieldValidation } from '@presentation/helpers/validators/required-field-validation'
-import { Validation } from '@presentation/helpers/validators/validation'
+import { Validation } from '@presentation/protocols/validation'
 import { ValidationComposite } from '@presentation/helpers/validators/validation-composite'
 import { EmailValidatorAdapter } from '@presentation/utils/email-validator-adapter'
