+++ src/main/factories/signup/signup-validation.spec.ts
@@ -1,7 +1,7 @@
 import { CompareFieldsValidation } from '@presentation/helpers/validators/compare-fields-validation'
 import { EmailValidation } from '@presentation/helpers/validators/email-validation'
 import { RequiredFieldValidation } from '@presentation/helpers/validators/required-field-validation'
-import { Validation } from '@presentation/helpers/validators/validation'
+import { Validation } from '@presentation/protocols/validation'
 import { ValidationComposite } from '@presentation/helpers/validators/validation-composite'
 import { EmailValidator } from '@presentation/protocols/email-validator.interface'
 import { makeSignUpValidation } from './signup-validation'
