+++ src/presentation/controller/login/login-protocols.ts
@@ -1,3 +1,3 @@
 export * from '@presentation/protocols' // import generic protocols for controllers
 export * from '@domain/usecases/authentication';
-export * from '@presentation/helpers/validators/validation'
+export * from '@presentation/protocols/validation'
