+++ src/presentation/controller/signup/signup-protocols.ts
@@ -1,4 +1,4 @@
 export * from '@presentation/protocols' // import generic protocols for controllers
 export * from '@domain/usecases/add-account';
 export * from '@domain/models/account';
-export * from '@presentation/helpers/validators/validation';
+export * from '@presentation/protocols/validation';
