+++ src/presentation/helpers/validators/compare-fields-validation.ts
@@ -1,5 +1,5 @@
 import { InvalidParamError } from '@presentation/errors'
-import { Validation } from './validation'
+import { Validation } from '../../protocols/validation'
 
 export class CompareFieldsValidation implements Validation {
   private readonly fieldName: string
