+++ src/presentation/helpers/validators/email-validation.ts
@@ -1,6 +1,6 @@
 import { InvalidParamError } from '@presentation/errors'
 import { EmailValidator } from '@presentation/protocols/email-validator.interface'
-import { Validation } from './validation'
+import { Validation } from '../../protocols/validation'
 
 export class EmailValidation implements Validation {
   private readonly fieldName: string
