+++ src/presentation/helpers/validators/validation-composite.spec.ts
@@ -1,5 +1,5 @@
 import { MissingParamError } from '@presentation/errors'
-import { Validation } from './validation'
+import { Validation } from '../../protocols/validation'
 import { ValidationComposite } from './validation-composite'
 
 const makeValidation = (): Validation => {
